# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: Analysis public REST API
  version: 0.0.2
paths:
  /analyses:
    summary: 'Manage analyses'
    post:
      summary: 'Create an analysis'
      description: ""
      operationId: createAnalysisV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisDTO'
      responses:
        '201':
          description: 'Analysis created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDTO'
        '400':
          description: 'Invalid input'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      tags:
        - 'Analysis'
    get:
      summary: 'Get a list of analysis of the logged user'
      description: ""
      operationId: listAnalysesV1
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAnalyses'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: 'Default error sample response'
      tags:
        - 'Analysis'
  /analyses/{analysisId}:
    summary: 'Manage an analysis'
    parameters:
      -
        name: analysisId
        in: path
        required: true
        description: 'The uid of the analysis to retrieve'
        schema:
          type: string
    get:
      summary: 'Get an analysis'
      description: ""
      operationId: getAnalysisV1
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDTO'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/UnaccessibleError'
        '404':
          $ref: '#/components/responses/AnalysisNotFoundError'
      tags:
        - 'Analysis'
    patch:
      summary: 'Update an analysis (Only status and visibility are updatable)'
      description: ""
      operationId: patchAnalysisV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisUpdatableDTO'
      responses:
        '204':
          description: "Analysis updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDTO'
        '400':
          $ref: '#/components/responses/InvalidStatusError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/UnaccessibleError'
        '404':
          $ref: '#/components/responses/AnalysisNotFoundError'
      tags:
        - 'Analysis'
    delete:
      summary: "Set the status of an analysis to cancelled"
      operationId: deleteAnalysisV1
      responses:
        '200':
          description: "Status change scheduled"
        '400':
          $ref: '#/components/responses/InvalidStatusError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/UnaccessibleError'
        '404':
          $ref: '#/components/responses/AnalysisNotFoundError'
      tags:
        - 'Analysis'
  /analyses/{analysisId}/history:
    summary: 'Get the status history of an analysis'
    parameters:
      -
        name: analysisId
        in: path
        required: true
        description: 'The uid of the analysis to retrieve'
        schema:
          type: string
    get:
      summary: 'Get the status history of an analysis'
      description: ""
      operationId: getAnalysisHistoryV1
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisStatusHistoryDTO'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/UnaccessibleError'
        '404':
          $ref: '#/components/responses/AnalysisNotFoundError'
      tags:
        - 'Analysis'
...
  /documents:
    summary: ""
    post:
      summary: "Upload a new document"
      description: ""
      operationId: uploadDocumentV1
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
      responses:
        '201':
          description: "File uploaded"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      tags:
        - 'Documents'
  /documents/{documentId}:
    summary: ""
    parameters:
      - name: documentId
        in: path
        required: true
        description: 'The uid of the document'
        schema:
          type: string
    get:
      summary: "Get document info"
      description: ""
      operationId: getDocumentMetaV1
      responses:
        '200':
          description: "The document specs"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/DocumentNotFoundError'
      tags:
        - 'Documents'
  /documents/{documentId}/download:
    summary: ""
    parameters:
      - name: documentId
        in: path
        required: true
        description: 'The uid of the document'
        schema:
          type: string
    get:
      summary: "Download document content"
      description: ""
      operationId: downloadDocumentV1
      responses:
        '200':
          description: "The content of the document"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/UnaccessibleError'
        '404':
          $ref: '#/components/responses/DocumentNotFoundError'
      tags:
        - 'Documents'
components:
  schemas:
    PagedResponse:
      type: object
      properties:
        totalCount:
          type: integer
          format: int64
        pageSize:
          type: integer
        page:
          type: integer
        count:
          type: integer
    UserDTO:
      type: object
      properties:
        uid:
          type: string
        username:
          type: string
    AnalysisStatusEnum:
      type: string
      default: ready
      enum:
        - ready
        - started
        - stopped
        - completed
        - cancelled
        - failed
    AnalysisVisibilityEnum:
      type: string
      default: public
      enum:
        - private
        - public
    AnalysisTypeEnum:
      type: string
      enum:
        - twitter-neel
    AnalysisInputTypeEnum:
      type: string
      enum:
        - query
        - dataset
    AnalysisInputDTO:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/AnalysisInputTypeEnum'
    QueryAnalysisInputDTO:
      allOf:
        - $ref: '#/components/schemas/AnalysisInputDTO'
        - type: object
          required:
            - tokens
            - operator
          properties:
            type:
              $ref: '#/components/schemas/AnalysisInputTypeEnum'
            tokens:
              type: array
              items:
                type: string
            joinOperator:
              type: string
              enum:
                - all
                - any
    DatasetAnalysisInputDTO:
      allOf:
        - $ref: '#/components/schemas/AnalysisInputDTO'
        - type: object
          required:
            - documentId
          properties:
            type:
              $ref: '#/components/schemas/AnalysisInputTypeEnum'
            documentId:
              type: string
    AnalysisDTO:
      type: object
      required:
        - type
        - input
      properties:
        id:
          type: string
          readOnly: true
        type:
          $ref: '#/components/schemas/AnalysisTypeEnum'
        owner:
          allOf:
            - $ref: '#/components/schemas/UserDTO'
            - readOnly: true
        status:
          $ref: '#/components/schemas/AnalysisStatusEnum'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisStatusHistoryDTO'
        visibility:
          $ref: '#/components/schemas/AnalysisVisibilityEnum'
        input:
          type: object
          oneOf:
            - $ref: '#/components/schemas/QueryAnalysisInputDTO'
            - $ref: '#/components/schemas/DatasetAnalysisInputDTO'
          discriminator:
            propertyName: type
            mapping:
              query: '#/components/schemas/QueryAnalysisInputDTO'
              dataset: '#/components/schemas/DatasetAnalysisInputDTO'
        progress:
          type: number
          format: double
          description: "A negative value if the input is an endless streaming, a value between [0, 1] otherwise"
          readOnly: true
        resultsCount:
          type: number
          format: int64
          description: "The number of results available for this analysis. While started, this value can be outdated."
          readOnly: true
        createDate:
          type: string
          format: date-time
          readOnly: true
        updateDate:
          type: string
          format: date-time
          readOnly: true
    AnalysisUpdatableDTO:
      properties:
        status:
          $ref: '#/components/schemas/AnalysisStatusEnum'
        visibility:
          $ref: '#/components/schemas/AnalysisVisibilityEnum'
    AnalysisStatusHistoryDTO:
      type: object
      properties:
        oldStatus:
          $ref: '#/components/schemas/AnalysisStatusEnum'
        newStatus:
          $ref: '#/components/schemas/AnalysisStatusEnum'
        message:
          type: string
        errorCode:
          type: integer
        user:
          $ref: '#/components/schemas/UserDTO'
        date:
          type: string
          format: date-time
    PagedAnalyses:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            objects:
              type: array
              items:
                $ref: '#/components/schemas/AnalysisDTO'
...
    DocumentDTO:
      type: object
      properties:
        documentId:
          type: string
        filename:
          type: string
        size:
          type: integer
          format: int64
        contentType:
          type: string
        uploadDate:
          type: string
          format: 'date-time'
        user:
          $ref: '#/components/schemas/UserDTO'
  parameters:
    pageSizeParam:
      name: pageSize
      in: query
      description: 'How many items to return per page'
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 2000
        default: 100
    pageParam:
      name: page
      in: query
      description: 'The page from which start to fetch items (zero-based)'
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
  responses:
    UnauthorizedError:
      description: JWT token is missing or invalid
    UnaccessibleError:
      description: Cannot access to this resource
    AnalysisNotFoundError:
      description: Analysis not found
    InvalidStatusError:
      description: Analysis status change cannot be applied
    DocumentNotFoundError:
      description: Document not found
servers:
  -
    url: /api/public
    variables: {}
    description: 'Analysis microservice'